+++ packages/SUI.CORE/src/user.ts
@@ -1,4 +1,4 @@
-export interface IUser {
+export interface ICoreUser {
   /**
    *   User Bearer access token
    */