+++ packages/SUI.META/src/utils/init.ts
@@ -1,7 +1,8 @@
-import { IBaseCardTabLayout } from '@smsoft/sui-base-components/src/Base/BaseCardTabLayout';
-import { IUser } from '@smsoft/sui-core';
+import { IBaseCardTabLayout } from '@smsoft/sui-base-components';
+import { ICoreUser } from '@smsoft/sui-core';
 
 import { ColumnInfo, ColumnInfoManager, NameManager, TableInfo, TableInfoManager } from '../cache';
+import { TableSettings } from '../TableSettings';
 
 import { parseRoutes, RouteType } from './metaUtils';
 
@@ -36,13 +37,13 @@ export interface IRawRoute {
 }
 
 export interface IMetaInitProps {
-  // Tmp
-  // tslint:disable-next-line:no-any
-  extraOmniTableSettingsTab?: IBaseCardTabLayout<any>;
   routes?: IRawRoute[];
-  user?: IUser,
+  user?: ICoreUser,
 
   defaultGetLinkForTable?(tableName: string, type: RouteType, id?: string | number): string | null;
+  // Tmp
+  // tslint:disable-next-line:no-any variable-name
+  extraOmniTableSettingsTab?(_this: TableSettings): IBaseCardTabLayout<any>;
   // tslint:disable-next-line:no-any
   routerPushFn(link: any): void;
 }